incorporate requested changes partially
