+++ requirements.txt
@@ -6,7 +6,7 @@ moscot
 networkx>=2.2
 numba>=0.51.0
 numpy>=1.17.0
-pandas==1.5.0
+pandas<2
 pygam>=0.8.0
 pygpcca>=1.0.3
 scanpy>=1.7.2
